requirements.txt
anyio==4.8.0
APScheduler==3.11.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
